+++ appendIfNotExists.js
