+++ removeIfExists.js
